requirements.txt
streamlit>=1.28.0
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0
peft>=0.6.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
Pillow>=10.0.0
python-pptx>=0.6.23
numpy>=1.24.0
colour>=0.1.5

